refactor(spi): extract policy SPIs into com.embabel.dice.spi and fix common layering
Move the policy extension points — TrustScorer, AuthorityResolver/AuthorityTier,
AuthorityWeightedTrustScorer, ConflictDetector, and ConflictType (plus their
shipped defaults) — out of common and proposition.revision into a dedicated
com.embabel.dice.spi package, giving the library's extension points one home.

This also fixes a layering wrinkle: common no longer imports
proposition.revision.ConflictType. ConflictType now lives in spi alongside the
policies that share it, and proposition.revision depends downward on spi.

Tests mirror the move into the spi package; no behaviour change.
