+++ dice/src/main/kotlin/com/embabel/dice/common/DiceMetadataKeys.kt
@@ -24,7 +24,7 @@ package com.embabel.dice.common
 object DiceMetadataKeys {
 
     /**
-     * Cached trust score for a proposition — a `Double` in `[0.0, 1.0]` produced by a [TrustScorer].
+     * Cached trust score for a proposition — a `Double` in `[0.0, 1.0]` produced by a `TrustScorer`.
      *
      * It's a cache: a content revision (`withText` / `withConfidence`) re-anchors the decay clock but
      * does not clear this key, so the score can briefly outlive the content it was computed for. That's
