+++ dice/src/main/kotlin/com/embabel/dice/proposition/revision/PropositionReviser.kt
@@ -17,6 +17,7 @@ package com.embabel.dice.proposition.revision
 
 import com.embabel.dice.proposition.Proposition
 import com.embabel.dice.proposition.PropositionRepository
+import com.embabel.dice.spi.ConflictType
 import com.fasterxml.jackson.annotation.JsonPropertyDescription
 
 /**
