+++ dice/src/main/kotlin/com/embabel/dice/spi/ConflictType.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.embabel.dice.proposition.revision
+package com.embabel.dice.spi
 
 /**
  * Classifies the nature of a conflict between an incoming proposition and an
