+++ dice/src/test/kotlin/com/embabel/dice/spi/ConflictDetectorTest.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.embabel.dice.proposition.revision
+package com.embabel.dice.spi
 
 import com.embabel.agent.core.ContextId
 import com.embabel.dice.proposition.Proposition
