+++ dice/src/test/kotlin/com/embabel/dice/spi/TrustScorerTest.kt
@@ -13,11 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.embabel.dice.common
+package com.embabel.dice.spi
 
 import com.embabel.agent.core.ContextId
 import com.embabel.dice.proposition.Proposition
-import com.embabel.dice.proposition.revision.ConflictType
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
